Moved js tests to __tests__ directory and removed 'use strict' from js files
