+++ package.json
@@ -11,7 +11,7 @@
     "build": "gulp build",
     "test": "npm run test-crtool && npm run test-js",
     "test-crtool": "gulp test",
-    "test-js": "jest teachers_digital_platform/js --coverage",
+    "test-js": "jest teachers_digital_platform/__tests__ --coverage",
     "watch": "gulp watch",
     "postinstall": "cd teachers_digital_platform/crtool && npm install"
   },
