+++ teachers_digital_platform/js/expandable-facets.js
@@ -2,8 +2,6 @@
    Expandable Facets Organism
    ========================================================================== */
 
-'use strict';
-
 // polyfill for ie9 compatibility
 require( 'classlist-polyfill' );
