+++ teachers_digital_platform/js/tdp-utils.js
@@ -1,5 +1,3 @@
-'use strict';
-
 const xhr = require( './util/ajax-request' ).ajaxRequest;
 
 /**
