+++ teachers_digital_platform/js/util/EventObserver.js
@@ -1,5 +1,3 @@
-'use strict';
-
 /**
  * EventObserver
  * @class
