+++ teachers_digital_platform/js/util/behavior.js
@@ -1,5 +1,3 @@
-'use strict';
-
 
 /* ==========================================================================
    Dom Behaviors
