+++ teachers_digital_platform/js/util/data-hook.js
@@ -1,5 +1,3 @@
-'use strict';
-
 
 // Required modules.
 const standardType = require( './standard-type' );
