+++ teachers_digital_platform/js/util/dom-traverse.js
@@ -1,5 +1,3 @@
-'use strict';
-
 
 const typeCheckers = require( './type-checkers' );
