+++ teachers_digital_platform/js/util/standard-type.js
@@ -1,5 +1,3 @@
-'use strict';
-
 
 /**
  * @constant
