+++ teachers_digital_platform/js/util/type-checkers.js
@@ -1,5 +1,3 @@
-'use strict';
-
 
 /* ==========================================================================
    Javascript Type Checkers
